+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- [#726]: `defmt-decoder`: Remove difference in favor of dissimilar
 - [#715]: Update `clap` to version `4`
 - [#710]: Update CI
 - [#706]: Satisfy clippy
@@ -28,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [#678]: Satisfy clippy
 - [#719]: defmt-print: fix panic 
 
+[#726]: https://github.com/knurling-rs/defmt/pull/726
 [#715]: https://github.com/knurling-rs/defmt/pull/715
 [#710]: https://github.com/knurling-rs/defmt/pull/710
 [#706]: https://github.com/knurling-rs/defmt/pull/706